Fixed all GCC warnings when bootstrapped in 'pedantic' mode except one, and it's only because I don't understand what it's babbling about.
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@2031 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ examples/fieldinf.cpp
@@ -63,7 +63,7 @@ main(int argc, char *argv[])
 				setw(widths[1]) << "SQL Type" <<
 				setw(widths[2]) << "Equivalent C++ Type" <<
 				endl;
-		for (int i = 0; i < sizeof(widths) / sizeof(widths[0]); ++i) {
+		for (size_t i = 0; i < sizeof(widths) / sizeof(widths[0]); ++i) {
 			cout << string(widths[i] - 1, '=') << ' ';
 		}
 		cout << endl;
